Refactor acceptance tests
Moved several definitions into common and made them more general purpose
