+++ acj/static/test/features/create_course.feature
@@ -4,13 +4,13 @@ Feature: Create Course
   Scenario: Loading add course page by Add Course button as admin
     Given I'm "admin"
     And I'm on "home" page
-    When I select 'Add Course' button
+    When I select "Add Course" button
     Then "Add Course" page should load
 
   Scenario: Loading add course page by add a course button as instructor
     Given I'm "instructor1"
     And I'm on "home" page
-    When I select 'Add Course' button
+    When I select "Add Course" button
     Then "Add Course" page should load
 
   Scenario: Creating a course as instructor
@@ -19,6 +19,6 @@ Feature: Create Course
     And I fill in:
       | element     | content     |
       | course.name | Test Course |
-    When I click on "Save" button
+    When I submit form with "Save" button
     Then I should be on "course" page
     And I should see "Test Course" in "h1" on the page
