+++ acj/static/test/page_objects/create_course.js
@@ -9,10 +9,15 @@ var CreateCoursePage = function() {
     this.get = function() {
         return browser.get(env.baseUrl + '#/course/new');
     };
-
-    this.addCourse = function() {
-        return addCourseButton.click();
-    };
+    
+    this.clickButton = function(button) {
+        switch (button) {
+            case "Add Course":
+                return addCourseButton.click();
+            case "Create User":
+                return createUserButton.click();
+        }
+    }
 };
 
 module.exports = CreateCoursePage;
